Finished new commands system, code cleanup
