+++ src/main/java/com/gamesense/api/config/SaveConfig.java
@@ -15,7 +15,7 @@ import com.gamesense.api.util.players.friends.Friend;
 import com.gamesense.api.util.players.friends.Friends;
 import com.gamesense.client.GameSenseMod;
 import com.gamesense.client.clickgui.GameSenseGUI;
-import com.gamesense.client.commands2.Command;
+import com.gamesense.client.command.Command;
 import com.gamesense.client.module.Module;
 import com.gamesense.client.module.ModuleManager;
 import com.gamesense.client.module.modules.misc.AutoGG;
@@ -44,7 +44,7 @@ public class SaveConfig {
         }
     }
 
-    String fileName = "GameSense/";
+    public static final String fileName = "GameSense/";
     String moduleName = "Modules/";
     String mainName = "Main/";
     String miscName = "Misc/";
