+++ src/main/java/com/gamesense/api/event/EventProcessor.java
@@ -5,8 +5,8 @@ import java.net.URL;
 import java.util.Map;
 
 import com.gamesense.api.util.render.GSColor;
-import com.gamesense.client.commands2.Command;
-import com.gamesense.client.commands2.CommandManager;
+import com.gamesense.client.command.Command;
+import com.gamesense.client.command.CommandManager;
 import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.modules.render.SkyColor;
 import net.minecraftforge.client.event.*;
