+++ src/main/java/com/gamesense/api/util/font/CFont.java
@@ -150,7 +150,7 @@ public class CFont{
 		tex = setupTexture(font, this.antiAlias, this.fractionalMetrics, this.charData);
 	}
 
-	protected class CharData{
+	protected static class CharData{
 		public int width;
 		public int height;
 		public int storedX;
