+++ src/main/java/com/gamesense/client/command/commands/FriendCommand.java
@@ -1,44 +1,35 @@
 package com.gamesense.client.command.commands;
 
-import com.gamesense.client.GameSenseMod;
 import com.gamesense.api.util.misc.MessageBus;
-import com.mojang.realmsclient.gui.ChatFormatting;
-import com.gamesense.client.command.Command;
 import com.gamesense.api.util.players.friends.Friends;
+import com.gamesense.client.command.Command;
 
-public class FriendCommand extends Command{
+/**
+ * @Author Hoosiers on 11/05/2020
+ */
 
-	@Override
-	public String[] getAlias(){
-		return new String[]{"friend", "friends", "f"};
-	}
+public class FriendCommand extends Command {
 
-	@Override
-	public String getSyntax(){
-		return "friend <add | del> <name>";
-	}
+    public FriendCommand(){
+        super("Friend");
 
-	@Override
-	public void onCommand(String command, String[] args) throws Exception{
-		if (args[0].equalsIgnoreCase("add")){
-			if (Friends.isFriend(args[1])){
-				MessageBus.sendClientPrefixMessage(args[1] + ChatFormatting.GRAY + " is already a friend!");
-				return;
-			}
-			if (!Friends.isFriend(args[1])){
-				GameSenseMod.getInstance().friends.addFriend(args[1]);
-				MessageBus.sendClientPrefixMessage("Added " + args[1] + " to friends list");
-			}
-		}
-		if (args[0].equalsIgnoreCase("del") || args[0].equalsIgnoreCase("remove")){
-			if (!Friends.isFriend(args[1])){
-				MessageBus.sendClientPrefixMessage(args[1] + " is not a friend!");
-				return;
-			}
-			if (Friends.isFriend(args[1])){
-				GameSenseMod.getInstance().friends.delFriend(args[1]);
-				MessageBus.sendClientPrefixMessage("Removed " + args[1] + " from friends list");
-			}
-		}
-	}
+        setCommandSyntax(Command.getCommandPrefix() + "friend add/del [player]");
+        setCommandAlias(new String[]{
+                "friend", "friends", "f"
+        });
+    }
+
+    public void onCommand(String command, String[] message) throws Exception{
+        String main = message[0];
+        String value = message[1];
+
+        if (main.equalsIgnoreCase("add") && !Friends.isFriend(value)){
+            Friends.addFriend(value);
+            MessageBus.sendClientPrefixMessage("Added friend: " + value.toUpperCase() + "!");
+        }
+        else if (main.equalsIgnoreCase("del") && Friends.isFriend(value)){
+            Friends.addFriend(value);
+            MessageBus.sendClientPrefixMessage("Deleted friend: " + value.toUpperCase() + "!");
+        }
+    }
 }
