--- src/main/java/com/gamesense/client/command/commands/SaveConfigCommand.java
@@ -1,24 +1,0 @@
-package com.gamesense.client.command.commands;
-
-import com.gamesense.api.config.ConfigStopper;
-import com.gamesense.client.command.Command;
-import com.gamesense.api.util.misc.MessageBus;
-
-public class SaveConfigCommand extends Command{
-
-	@Override
-	public String[] getAlias(){
-		return new String[]{"saveconfig", "savecfg"};
-	}
-
-	@Override
-	public String getSyntax(){
-		return "saveconfig";
-	}
-
-	@Override
-	public void onCommand(String command, String[] args) throws Exception{
-		ConfigStopper.saveConfig();
-		MessageBus.sendClientPrefixMessage("Config saved!");
-	}
-}
